+++ src/app/@shared/services/project.service.ts
@@ -1,25 +1,46 @@
 import { Injectable } from '@angular/core';
-import {HttpClient, HttpHeaders} from '@angular/common/http';
+import { HttpClient, HttpHeaders , HttpEventType } from '@angular/common/http';
+import { environment } from '../../../environments/environment';
+import { map } from  'rxjs/operators';
 
 @Injectable()
 export class ProjectService {
-
+    baseUrl: string;
     projects;
-    constructor(private http: HttpClient) { }
+
+    constructor(private http: HttpClient) {
+        this.baseUrl = `${environment.apiEndpooint}/projects`;
+    }
 
     async getProjects() {
-
-        const url = "http://localhost:9090/projects"
-        let asyncResult = await this.http.get(url, this.httpOptions()).toPromise();
-        console.log('ProjectService getProjects | ', asyncResult);
+        let asyncResult = await this.http.get(this.baseUrl, this.httpOptions()).toPromise();
         this.projects = asyncResult;
         return asyncResult
     }
 
-    private httpOptions() {
+    async uploadDeliveryReport(projectId, data, successCb, errorCb) {
+        const url = `${this.baseUrl}/${projectId}/deliveryReports/upload`;
+        return await this.http.post<any>(url, data, {
+            reportProgress: true,
+            observe: 'events'
+        }).pipe(map((event) => {
+              switch (event.type) {
+                  case HttpEventType.UploadProgress:
+                      const progress = Math.round(100 * event.loaded / event.total);
+                      return { status: 'progress', message: progress };
+                  case HttpEventType.Response:
+                      return {
+                        deliveryReport: event.body
+                      };
+                  default:
+                      return `Unhandled event: ${event.type}`;
+              }
+        })).subscribe(successCb, errorCb);
+    }
 
-        return {
-          headers: new HttpHeaders({ 'Content-Type':  'application/json'})
-        }
+    private httpOptions() {
+      return {
+        headers: new HttpHeaders({ 'Content-Type':  'application/json'})
       }
+    }
 }
