+++ src/app/pages/delivery-reports/delivery-reports.component.scss
@@ -32,10 +32,17 @@
 }
 
 .status {
-    font-size: 20px !important;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
     padding: 0.5rem;
+}
+
+.status-icon {
+    font-size: 20px;
+}
+
+.overall-status-row {
+    border-top: 1px solid rgba(0, 0, 0, 0.12);
 }
