+++ src/environments/environment.prod.ts
@@ -1,4 +1,6 @@
 
 export const environment = {
   production: true,
+  // Change when production env. known
+  apiEndpooint: 'http://localhost:9000',
 };
