+++ src/environments/environment.ts
@@ -1,4 +1,5 @@
 
 export const environment = {
   production: false,
+  apiEndpooint: 'http://localhost:9090',
 };
